Move strides utilities to `strides.h`
